all menus except in-game menu and player inventory
In the scope of Menus, everything but the in-game menu and player inventory has been implemented through scripts following the project's MVC architecture.

Side note: Upon loading the Menu scene, there shouldn't be anything present in the editor. This is deliberately done as the scripts handle all GUIs to be rendered. Therefore, running the scene will render the GUIs.